Change default to run asort() on the array [touch: 6651]
I discussed this with Sidney in chat and now I'm thinking it'd be best to fix this using the asort function by default. The filter can be used to disable sorting entirely for that edge case where this just works, and if their categories are in a different order, the filter can be used to reverse sort as well.